Moved contents of run_gui.py to various app files
